backend/requirements.txt
boto3==1.26.112
cryptography==40.0.1
python-dotenv==1.0.0
